Zarr: split ZarrArray in 2 classes (no functional change)
