+++ frmts/zarr/CMakeLists.txt
@@ -1,6 +1,8 @@
 add_gdal_driver(
   TARGET gdal_Zarr
   SOURCES zarr_array.cpp
+          zarr_v2_array.cpp
+          zarr_v3_array.cpp
           zarr_attribute.cpp
           zarr_dimension.cpp
           zarr_group.cpp
